+++ R/Design.R
@@ -48,7 +48,8 @@ Design = R6Class("Design",
 
     #' @description
     #' Helper for print outputs.
-    format = function() {
+    #' @param ... (ignored).
+    format = function(...) {
       sprintf("<%s>", class(self)[1L])
     },