+++ tests/testthat/test_to_tune.R
@@ -83,7 +83,6 @@ test_that("$check() works on TuneToken", {
   expect_error(pars$search_space(list(lgl = to_tune(0, 1))), "must have zero or one argument")
 
   expect_error(pars$search_space(list(xxx = to_tune())), "ust be a subset of .*x.*xub.*y.*uty.*uty1.*fct.*lgl")
-
 })
 
 test_that("Tune ParamSet is created", {