+++ NEWS.md
@@ -1,17 +1,38 @@
-
-# paradox 0.8.0
+# paradox 0.12.0
 * Removed `Param` objects. `ParamSet` now uses a `data.table` internally; individual parameters are more like `Domain` objects now. `ParamSets` should be constructed using the `ps()` shorthand and `Domain` objects. This entails the following major changes:
     * `ParamSet` now supports `extra_trafo` natively; it behaves like `.extra_trafo` of the `ps()` call.
     * `ParamSet` has `$constraint`
     * `ParamSet` objects are now less mutable. The only properties that can be changed are `values`, `tags`, `deps`, `constraint` and `extra_trafo`.
     * `ParamSet$is_bounded` is a vector with an entry for each parameter. Use `$all_bounded` for the previous behavior.
     * `Condition` objects are now S3 objects and can be constructed with `CondEqual()` and `CondAnyOf()`, instead of `CondXyz$new()`. (It is recommended to use the `Domain` interface for conditions, which has not changed)
+* Added a vignette which was previously a chapter in the `mlr3book`
 
+# paradox 0.11.1
 
-# paradox 0.7.1.9000
+* Minor bug fixes.
 
-- Same as previous version.
+# paradox 0.11.0
 
+* feat: The function `generate_design_sobol()` generates a space-filling Sobol sequence design.
+* refactor: `$set_values` returns the parameter set invisible.
+
+# paradox 0.10.0
+
+* Reset `.has_extra_trafo` to `FALSE` when trafo is set to `NULL`.
+* `rd_info.ParamSet` collapses vector with `"\n"` due changes in roxygen 7.2.0
+* Add method `set_values()` to conveniently add parameter values.
+
+# paradox 0.9.0
+
+* Added `default_values()` function to extract default values from `ParamSet`
+  objects.
+
+# paradox 0.8.0
+
+* Parameters now have a new (optional) field `description`.
+* Improved printing of parameters in documentation (#355).
+* A warning is now signaled if the package `ParamHelpers` is also loaded.
+* Fixed some links.
 
 # paradox 0.7.1